svg/animations/animateTransform-translate-invalid-attributetype.html

<!doctype html>
<html>
<meta charset="utf-8">
<title>Tests that an animateTransform with invalid attributeType does not animate. Should result in a 100x100 rect and only PASS messages.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/SVGAnimationTestCase-testharness.js"></script>

<svg>
</svg>

<script>
var rootSVGElement = document.querySelector("svg");
var epsilon = 1.0;

// Setup test document

var rect = createSVGElement("rect");
rect.setAttribute("id", "rect");
rect.setAttribute("width", "200");
rect.setAttribute("height", "200");
rect.setAttribute("fill", "green");
rect.setAttribute("onclick", "executeTest()");

var animate = createSVGElement("animateTransform");
animate.setAttribute("id", "animation");
animate.setAttribute("attributeName", "transform");
animate.setAttribute("attributeType", "CSS");
animate.setAttribute("type", "translate");
animate.setAttribute("from", "0, 0");
animate.setAttribute("to", "200, 200");
animate.setAttribute("begin", "0s");
animate.setAttribute("dur", "4s");
animate.setAttribute("fill", "freeze");
rect.appendChild(animate);

rootSVGElement.appendChild(rect);

// Setup animation test
function sample() {
    let translation = rect.getCTM().inverse().multiply(rootSVGElement.getCTM());
    assert_approx_equals(translation.e, 0, epsilon);
    assert_approx_equals(translation.f, 0, epsilon);
}

smil_async_test((t) => {
    const expectedValues = [
        // [animationId, time, sampleCallback]
        ["animation", 0.0,   sample],
        ["animation", 0.001, sample],
        ["animation", 2.0,   sample],
        ["animation", 3.999, sample],
        ["animation", 4.001, sample]
    ];

    runAnimationTest(t, expectedValues);
});

</script>
